ref(dynamic-sampling): co-locate schema in organizationSampling
Move targetSampleRateSchema from standalone utils file into
organizationSampling.tsx to match codebase convention of defining
schemas inline. The modal imports it from there.
